requirements.txt
bcrypt==4.0.1
boto3==1.28.62
botocore==1.31.62
click==7.1.2
Flask==1.1.4
Flask-Bcrypt==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==1.0.1
MarkupSafe==2.0.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.7.0
six==1.16.0
urllib3==1.26.17
Werkzeug==1.0.1
